Backend/requirements.txt
flask
azure-identity
openai
azure-appconfiguration-provider
featuremanagement